Decoupled FnF router
